rename to TargetDistributionModeSelect
